-Move logic to employee service.
